+++ PolicyManagement.Api/PolicyManagement.Models/Motor/PremiumFormDataModel.cs
@@ -7,26 +7,26 @@
         public int NonElectricAccessoriesIdv { get; set; }
         public int CngLpgIdv { get; set; }
         public int TotalIdv { get; set; }
-        public int Od { get; set; }
-        public int AddOnRiderOd { get; set; }
-        public int EndorseOd { get; set; }
+        public decimal Od { get; set; }
+        public decimal AddOnRiderOd { get; set; }
+        public decimal EndorseOd { get; set; }
         public int TotalOd { get; set; }
-        public int Tp { get; set; }
-        public int PassengerCover { get; set; }
-        public int EndorseTp { get; set; }
+        public decimal Tp { get; set; }
+        public decimal PassengerCover { get; set; }
+        public decimal EndorseTp { get; set; }
         public int TotalTp { get; set; }
-        public int NonCommissionComponentPremium { get; set; }
+        public decimal NonCommissionComponentPremium { get; set; }
         public decimal GstPercentage { get; set; }
         public decimal GstValue { get; set; }
         public decimal GrossPremium { get; set; }
-        public int EndorseGrossPremium { get; set; }
+        public decimal EndorseGrossPremium { get; set; }
         public int TotalGrossPremium { get; set; }
         public decimal SpecialDiscount { get; set; }
         public decimal Loading { get; set; }
         public short Ncb { get; set; }
         public short CommissionPaidOn { get; set; }
         public decimal CommissionablePremium { get; set; }
-        public int BasicTpGstPercentage { get; set; }
+        public decimal BasicTpGstPercentage { get; set; }
         public int NetPremium { get; set; }
     }
 }
